refactor: restructure into cargo workspace with murmur-core
Split the crate into a workspace with two members:
- murmur-core: audio capture, transcription, VAD, config, and context abstractions
- murmur: binary with app loop, UI, input handling, and platform integration
