+++ Cargo.toml
@@ -1,87 +1,13 @@
-[package]
-name = "murmur"
-version = "0.1.1"
+[workspace]
+members = ["crates/murmur-core", "crates/murmur"]
+resolver = "2"
+
+[workspace.package]
 edition = "2021"
-description = "Cross-platform, local voice dictation. Worry free speech to text. Powered by whisper.cpp, runs 100% on-device."
 license = "MIT"
 repository = "https://github.com/jacobfreck/murmur"
-readme = "README.md"
-keywords = ["whisper", "dictation", "voice", "speech-to-text"]
-categories = ["multimedia::audio"]
-
-[dependencies]
-# Speech-to-text: direct FFI bindings to whisper.cpp (statically linked)
-whisper-rs = "0.16"
-
-# Audio capture + WAV writing
-cpal = "0.17"
-hound = "3.5"
-
-# System tray icon + context menu (re-exports muda as tray_icon::menu)
-tray-icon = "0.21"
-
-# Global keyboard events (key-down + key-up for hold-to-talk)
-rdev = "0.5"
-
-# Clipboard access (by 1Password)
-arboard = "3"
-
-# Simulated keyboard input (Ctrl+V / Cmd+V paste)
-enigo = "0.6"
-
-# Config + serialization
-serde = { version = "1", features = ["derive"] }
-serde_json = "1"
-
-# Model download from HuggingFace
-reqwest = { version = "0.13", features = ["blocking"] }
-
-# Cross-platform config/data directories
-dirs = "6"
-
-# Text post-processing (spoken punctuation)
-regex = "1"
-
-# CLI argument parsing
-clap = { version = "4", features = ["derive"] }
-
-# Error handling
-anyhow = "1"
-
-# Voice activity detection (Silero VAD via ONNX Runtime)
-voice_activity_detector = "0.2"
-
-# Logging
-log = "0.4"
-env_logger = "0.11"
-
-# Noise suppression (pure Rust RNNoise port)
-nnnoiseless = "0.5"
-png = "0.18.1"
-
-# Transparent overlay window
-eframe = { version = "0.33", default-features = false, features = ["glow", "default_fonts", "x11", "wayland"] }
-egui = "0.33"
-
-# Timestamp formatting for note filenames
-chrono = "0.4"
-
-# macOS: Cocoa event loop for system tray (already transitive deps of tray-icon)
-[target.'cfg(target_os = "macos")'.dependencies]
-objc2-app-kit = { version = "0.3", features = ["NSApplication", "NSRunningApplication", "NSEvent", "NSWorkspace", "NSWindow"] }
-objc2-foundation = { version = "0.3", features = ["NSDate", "NSRunLoop"] }
-objc2 = "0.6"
-
-[features]
-default = []
-cuda = ["whisper-rs/cuda"]
-metal = ["whisper-rs/metal"]
-vulkan = ["whisper-rs/vulkan"]
 
 [profile.release]
 opt-level = 3
 lto = true
 strip = true
-
-[dev-dependencies]
-tempfile = "3"
