+++ crates/murmur-core/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "murmur-core"
+version = "0.1.1"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Core transcription engine for murmur — audio capture, Whisper transcription, VAD, and context abstractions."
+
+[dependencies]
+whisper-rs = "0.16"
+cpal = "0.17"
+hound = "3.5"
+arboard = "3"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+reqwest = { version = "0.13", features = ["blocking"] }
+dirs = "6"
+regex = "1"
+anyhow = "1"
+voice_activity_detector = "0.2"
+log = "0.4"
+nnnoiseless = "0.5"
+
+[features]
+default = []
+cuda = ["whisper-rs/cuda"]
+metal = ["whisper-rs/metal"]
+vulkan = ["whisper-rs/vulkan"]
+
+[dev-dependencies]
+tempfile = "3"
