+++ crates/murmur-core/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod audio;
+pub mod config;
+pub mod context;
+pub mod transcription;
