+++ crates/murmur/src/input/inserter.rs
