--- src/lib.rs
@@ -1,11 +1,0 @@
-pub mod app;
-pub mod audio;
-pub mod config;
-pub mod context;
-pub mod input;
-pub mod notes;
-pub mod platform;
-pub mod transcription;
-pub mod ui;
-
-pub const VERSION: &str = env!("CARGO_PKG_VERSION");
